对齐API文档，统一了响应包，去掉了无用查询

+++ backend/src/main/java/com/chinasoft/smokesensor/dto/AlarmRecordResponse.java
@@ -18,7 +18,7 @@ public class AlarmRecordResponse {
 
     private String alarmType;
 
-    private Double smokeValue;
+    private Integer smokeValue;
 
     private String riskLevel;
