+++ backend/src/main/java/com/chinasoft/smokesensor/dto/DeviceLatestDataResponse.java
@@ -18,7 +18,7 @@ public class DeviceLatestDataResponse {
 
     private LocalDateTime lastHeartbeat;
 
-    private Double currentSmokeValue;
+    private Integer currentSmokeValue;
 
     private String currentRiskLevel;
