+++ backend/src/main/java/com/chinasoft/smokesensor/dto/SensorDataResponse.java
@@ -14,7 +14,7 @@ public class SensorDataResponse {
 
     private String deviceId;
 
-    private Double smokeValue;
+    private Integer smokeValue;
 
     private String riskLevel;
